Possible fix for #39
I was able to reproduce the issue described in https://github.com/HenrikJoreteg/moonboots/issues/39
By restarting my Ampersand project (the sample project), and then refreshing the browser.  
By making the change in this PR, the error does not happen, and it appears that the app is working properly.
Caveat as big as a freight train: I am new to Ampersand, Browserify, and Moonboots ... so this PR is offered only as a suggestion, not an exhaustively tested Fix.